Updates to schema
Target for this is to update schema to be minimal and sensible to be implemented on the server. Additionally I have removed things that were hard or not possible on the server. Structure of the data is not flexible for some functionalities but we do not target here to build fully functional app. Just build an sandbox for us to test most of the graphql corner cases.